Add a parent build column and filter
To be able to filter failures by parent build number, e.g. filter all
failures for a-full build number xyz, add a --parent-build-filter (-p)
option to swattool. The option can be passed multiple times and can be a
regular expression.

It is also displayed by default in show-pending-failures, after the
column "Build" as "Parent".

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>
Co-developed-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>