all: remove github workflows
This was added very early in the project lifetime, while it was hosted
on my personal github. Now the project is hosted on Yocto project gits,
so the code style workflow is no longer running on each push.

Also, I recently added pre-commit hooks for ruff and mypy, they should
to the job.

Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>